pyproject.toml
[tool.poetry]
name = "screen-divider"
version = "0.1.0"
description = ""
authors = ["Divider <it-cio@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
yadisk = "^2.1.0"
python-dotenv = "^1.0.1"
pyqt5-qt5 = "5.15.2"
pyqt5 = "^5.15.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
